fix(cli): split openclaw hook tests, pin stdout, sort adapter imports
